RENAME the configuration file
