+++ tidoop-api/conf/tidoop-api.json
@@ -10,7 +10,7 @@
     "database": "cosmos"
   },
   "log": {
-    "file_name": "/var/log/tidoop/tidoop-mr-lib-api/tidoop-mr-lib-api.log",
+    "file_name": "/var/log/tidoop/tidoop-api/tidoop-api.log",
     "date_pattern": ".dd-MM-yyyy"
   }
 }
