+++ tidoop-api/src/logger.js
@@ -24,7 +24,7 @@
  */
 
 // Module dependencies
-var config = require('../conf/tidoop-mr-lib-api.json');
+var config = require('../conf/tidoop-api.json');
 var winston = require('winston');
 
 // Global variables
