+++ tidoop-api/src/tidoop_api.js
@@ -28,7 +28,7 @@ var Hapi = require('hapi');
 var boom = require('boom');
 var cmdRunner = require('./cmd_runner.js');
 var packageJson = require('../package.json');
-var config = require('../conf/tidoop-mr-lib-api.json');
+var config = require('../conf/tidoop-api.json');
 var mysqlDriver = require('./mysql_driver.js');
 var serverUtils = require('./server_utils.js');
 var logger = require('./logger.js');
